match scores are now 64 bits wide
Since cube values are 64 bit, scores and related variables must have the
same size.

+++ src/gibbon-game.c
@@ -44,7 +44,7 @@ struct _GibbonGamePrivate {
         gsize num_snapshots;
         GibbonGameSnapshot *snapshots;
 
-        gint score;
+        gint64 score;
         gboolean resigned;
 
         gboolean is_crawford;
